start/requirements.txt
azure-functions
requests
pillow
onnx
onnxruntime
numpy
opencv-python
